requirements.txt
tkinter == 8.5
PySimpleGUI == 4.60.5
flask == 2.3.2
mysql.connector == 2.2.9
